reorganisation and new theme 	modified:   .config/alacritty/alacritty.toml 	modified:   .config/alacritty/colours.toml 	deleted:    .config/fastfetch/config.jsonc 	deleted:    .config/fastfetch/json_schema.json 	deleted:    .config/fastfetch/nahida.sixel 	new file:   .config/rofi/background.png 	modified:   .config/rofi/colours.rasi 	typechange: .config/rofi/config.rasi 	deleted:    .config/rofi/images 	modified:   .config/sway/config.d/autostart 	new file:   .config/sway/config.d/colours 	typechange: .config/sway/config.d/decorations 	modified:   .config/sway/config.d/keybinds 	modified:   .config/sway/config.d/windowrules 	modified:   .config/sway/extra/scripts/pipewire.sh 	modified:   .config/sway/extra/scripts/themeswap.sh 	modified:   .config/sway/extra/scripts/wallpaper.sh 	deleted:    .config/sway/extra/themes/green/foot/colours.ini 	deleted:    .config/sway/extra/themes/green/rofi/config.rasi 	deleted:    .config/sway/extra/themes/green/sway/decorations 	deleted:    .config/sway/extra/themes/green/sway/hyprlock.conf 	deleted:    .config/sway/extra/themes/green/swaync/config.json 	deleted:    .config/sway/extra/themes/green/swaync/style.css 	renamed:    .config/sway/extra/themes/green/alacritty/colours.toml -> .config/sway/extra/themes/nahida/alacritty/colours.toml 	renamed:    .config/sway/extra/themes/green/fastfetch/config.jsonc -> .config/sway/extra/themes/nahida/fastfetch/config.jsonc 	renamed:    .config/sway/extra/themes/green/fastfetch/nahida.sixel -> .config/sway/extra/themes/nahida/fastfetch/nahida.sixel 	renamed:    .config/sway/extra/themes/green/rofi/images/nahida.png -> .config/sway/extra/themes/nahida/rofi/background.png 	renamed:    .config/sway/extra/themes/green/rofi/colours.rasi -> .config/sway/extra/themes/nahida/rofi/colours.rasi 	new file:   .config/sway/extra/themes/nahida/sway/colours 	new file:   .config/sway/extra/themes/nahida/swaync/colours.css 	renamed:    .config/sway/extra/themes/green/wallpaper.sh -> .config/sway/extra/themes/nahida/wallpaper.sh 	renamed:    .config/sway/extra/themes/green/waybar/colours.css -> .config/sway/extra/themes/nahida/waybar/colours.css 	new file:   ".config/sway/extra/themes/nahida/\350\212\261\347\245\236\350\257\236\347\245\255\345\212\250\346\200\201\345\243\201\347\272\270.mp4" 	renamed:    .config/sway/extra/themes/purple/alacritty/colours.toml -> .config/sway/extra/themes/omori/alacritty/colours.toml 	renamed:    .config/sway/extra/themes/purple/fastfetch/config.jsonc -> .config/sway/extra/themes/omori/fastfetch/config.jsonc 	renamed:    .config/sway/extra/themes/purple/fastfetch/mewo.txt -> .config/sway/extra/themes/omori/fastfetch/mewo.txt 	new file:   .config/sway/extra/themes/omori/omori.png 	new file:   .config/sway/extra/themes/omori/rofi/background.png 	renamed:    .config/sway/extra/themes/purple/rofi/colours.rasi -> .config/sway/extra/themes/omori/rofi/colours.rasi 	new file:   .config/sway/extra/themes/omori/sway/colours 	new file:   .config/sway/extra/themes/omori/swaync/colours.css 	new file:   .config/sway/extra/themes/omori/wallpaper.sh 	renamed:    .config/sway/extra/themes/purple/waybar/colours.css -> .config/sway/extra/themes/omori/waybar/colours.css 	deleted:    .config/sway/extra/themes/purple/foot/colours.ini 	deleted:    .config/sway/extra/themes/purple/rofi/config.rasi 	deleted:    .config/sway/extra/themes/purple/rofi/images/no-blur.png 	deleted:    .config/sway/extra/themes/purple/sway/decorations 	deleted:    .config/sway/extra/themes/purple/sway/hyprlock.conf 	deleted:    .config/sway/extra/themes/purple/swaync/config.json 	deleted:    .config/sway/extra/themes/purple/swaync/style.css 	deleted:    .config/sway/extra/themes/purple/wallpaper.sh 	new file:   .config/sway/extra/themes/tanger/alacritty/colours.toml 	new file:   .config/sway/extra/themes/tanger/rofi/background.png 	new file:   .config/sway/extra/themes/tanger/rofi/colours.rasi 	new file:   .config/sway/extra/themes/tanger/sway/colours 	new file:   .config/sway/extra/themes/tanger/swaync/colours.css 	new file:   .config/sway/extra/themes/tanger/tanger.png 	new file:   .config/sway/extra/themes/tanger/wallpaper.sh 	new file:   .config/sway/extra/themes/tanger/waybar/colours.css 	typechange: .config/sway/hyprlock.conf 	new file:   .config/swaync/colours.css 	typechange: .config/swaync/config.json 	modified:   .config/swaync/scripts/notification.sh 	typechange: .config/swaync/style.css 	modified:   .config/waybar/colours.css 	modified:   .config/waybar/config.jsonc 	modified:   .config/waybar/scripts/vpn-watch.sh 	modified:   .config/waybar/scripts/vpn.sh 	modified:   .config/waybar/style.css

+++ .config/sway/extra/scripts/pipewire.sh
@@ -1,17 +1,17 @@
 #!/usr/bin/env bash
 
-if ! grep -r systemd /sbin/init; then
-    pkill -x pipewire\|pipewire-pulse\|wireplumber
+if ! grep -r systemd /sbin/init \
+  && ! ls /usr/bin/ | grep openrc-init; then
+  pkill -x pipewire\|pipewire-pulse\|wireplumber
 
-    pidwait -x pipewire\|pipewire-pulse\|wireplumber
+  pidwait -x pipewire\|pipewire-pulse\|wireplumber
+  
+  pipewire &
+  pipewire-pulse &
 
-    pipewire &
-    pipewire-pulse &
-
-    sleep 1s
-    wireplumber &
+  sleep 1s
+  wireplumber &
+  notify-send pass -u critical
 fi
-
-sleep 0.2s
-
-paplay --volume=65535 $HOME/.config/sway/extra/assets/heal.wav &
+sleep 0.5s
+paplay --volume=65535 $HOME/.config/sway/extra/assets/startup.wav &
